Changed the request's log method to logFn
I don't much like this, but the updates to restify made this necessary. Even though restify
is just a strategy (not core), it was making it impossible to support the latest version of
restify. Strictly speaking, this should probably be a major bump but I'm not doing this to
v4 for just this.